auth/requirements.txt
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
gunicorn~=22.0
auth0-python~=4.7.0
python-jose~=3.3